+++ packages/dmn-js/CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to [dmn-js](https://github.com/bpmn-io/dmn-js) are documente
 
 ___Note:__ Yet to be released changes appear here._
 
+* `FEAT`: emit `saveXML` life-cycle events
+
 ## 5.0.0
 
 _Republished `v5.0.0-1` as stable version_.